+++ R/r6-logger.R
@@ -8,7 +8,7 @@ logger = R6::R6Class(
   public = list(
 
     start_logger = function() {
-      cli::cli_alert_info("Starting check: {private$context}")
+      cli::cli_alert_info("Starting check: {private$group}({private$short}): {private$context}")
 
       private$start_time = Sys.time()
     },