+++ NEWS.md
@@ -1,8 +1,10 @@
+# jrHealthCheckConnect 0.2.11 _2022-12-20_
+- feat: Advance quarto tests: python/Julia/Obser'
+
 # jrHealthCheckConnect 0.2.10 _2022-12-20_
 -  feat: quarto test: quarto to beamer
 -  fix: individual python tests now use ENV variables
 -  fix: Merge conflict broke deploy flask
-
 
 # jrHealthCheckConnect 0.2.9 _2022-11-07_
 -  feat: Perform a basic API call to check token